backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
dynaconf==3.2.4
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
minio==7.2.0
mongoengine==0.27.0
orjson==3.9.10
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
requests==2.28.2
requests-oauthlib==1.3.1
